refactor(agent): Refactor agents component
