+++ src/agents/func/recycle-instance-agents.ts
@@ -1,0 +1,50 @@
+import GameInstance from "../../game-instance";
+import { AgentReference, isAgentReference } from "../agent-reference";
+import { buildInstanceAgents } from "../impl/instance-agents-impl";
+import { InstanceAgents } from "../instance-agents";
+import { StaticAgentRegistry } from "../static-agent-registry";
+
+/**
+ * Creates an `InstanceAgents` for the new game cycle, keeping only
+ * the final properties of every agent from before.
+ *
+ * @param oldAgents     The `InstanceAgents` to recycle data from.
+ * @param newInstance   The new `GameInstance` that will own this `InstanceAgents`.
+ */
+export const recycleInstanceAgents = (
+    oldAgents: InstanceAgents,
+    newInstance: GameInstance
+): InstanceAgents => {
+    const newAgents = buildInstanceAgents(newInstance);
+    (newAgents as any)._nextId = (oldAgents as any)._nextId; // TODO: Fix
+
+    for (const formerAgent of oldAgents.agentManagers()) {
+        const id = formerAgent.id;
+        const am = newAgents.createAgentManager(id);
+
+        const propsToAdd = Object.keys(formerAgent).filter(
+            key => key !== "game" && key !== "id"
+        );
+
+        // For each updated property on the old agent, add its last value to the new agent.
+        propsToAdd.forEach(prop => {
+            if (formerAgent.propertyWasDeleted(prop)) {
+                if (StaticAgentRegistry.hasAgentProperty(id, prop)) {
+                    am.deleteProperty(prop); // Record deletions to static agents.
+                }
+
+                return; // If the property was deleted, don't add it to the new record.
+            }
+
+            let formerValue = formerAgent.getProperty(prop);
+
+            if (isAgentReference(formerValue)) {
+                formerValue = new AgentReference(formerValue.refId);
+            }
+
+            am.setProperty(prop, formerValue);
+        });
+    }
+
+    return newAgents;
+};
