+++ src/agents/impl/active-agent-proxy.ts
@@ -1,0 +1,70 @@
+import GameInstance from "../../game-instance";
+import { Agent } from "../agent-model";
+
+/** Builds the proxy handler for an active agent proxy. */
+const activeAgentProxyHandler = (id: number, game: GameInstance) => ({
+    get(target: Agent, property: PropertyKey) {
+        return game.agents.hasAgentProperty(id, property)
+            ? game.agents.getAgentProperty(id, property)
+            : Reflect.get(target, property);
+    },
+
+    set(target: Agent, property: PropertyKey, value: any) {
+        return game.agents.setAgentProperty(id, property, value);
+    },
+
+    has(target: Agent, property: PropertyKey) {
+        return game.agents.hasAgentProperty(id, property);
+    },
+
+    deleteProperty(target: Agent, property: PropertyKey) {
+        return game.agents.deleteAgentProperty(id, property);
+    },
+
+    getOwnPropertyDescriptor(target: Agent, property: PropertyKey) {
+        if (property === "length" && target instanceof Array) {
+            return Reflect.getOwnPropertyDescriptor(target, property);
+        } else {
+            return {
+                configurable: true,
+                enumerable: true,
+                value: this.get(target, property)
+            };
+        }
+    },
+
+    ownKeys(target: Agent) {
+        return game.agents.getAgentPropertyKeys(id);
+    },
+
+    getPrototypeOf(target: Agent) {
+        return Object.getPrototypeOf(target);
+    }
+});
+
+/**
+ * Builds a proxy for an active agent. When an inactive agent is activated
+ * by a `GameInstance`, it is considered active.
+ *
+ * The proxy wraps an empty object and has no tangible connection to the agent
+ * which it is imitating. All calls to the proxy are forwarded to the
+ * `GameInstance`'s `InstanceAgents`, simulating the behavior of normal object.
+ *
+ * @param id    The proxy agent's id.
+ * @param game  The `GameInstance` of the current context.
+ */
+export const activeAgentProxy = (id: number, game: GameInstance): Agent =>
+    new Proxy({} as any, activeAgentProxyHandler(id, game));
+
+/**
+ * Builds a proxy for an active agent array. An agent array is an array
+ * that is treated like an agent. All arrays that are properties of
+ * active agents become agent arrays.
+ *
+ * An agent array has all the same methods as a regular array.
+ *
+ * @param id    The agent array's id.
+ * @param game  The `GameInstance` of the current context.
+ */
+export const activeAgentArrayProxy = (id: number, game: GameInstance): Agent =>
+    new Proxy([] as any, activeAgentProxyHandler(id, game));
